Split todos model out
